DSEGOG-267 Move tests to `test/endpoints/`
