+++ test/endpoints/conftest.py
@@ -1,6 +1,9 @@
+import os
+
 import pytest_asyncio
 
 from operationsgateway_api.src.mongo.interface import MongoDBInterface
+from operationsgateway_api.src.records.echo_interface import EchoInterface
 
 
 async def add_user(auth_type):
@@ -55,3 +58,35 @@ async def delete_fed_fixture():
 async def delete_local_fixture():
     yield
     await remove_user("local")
+
+
+async def remove_record(timestamp_id):
+    await MongoDBInterface.delete_one(
+        "records",
+        filter_={"_id": f"{timestamp_id}"},
+    )
+
+
+async def remove_image(images):
+    echo = EchoInterface()
+    for image_path in images:
+        echo.delete_file_object(image_path)
+
+
+async def remove_waveform():
+    await MongoDBInterface.delete_one(
+        "waveforms",
+        filter_={"_id": "20200407142816_PM-201-HJ-PD"},
+    )
+
+
+@pytest_asyncio.fixture(scope="function")
+async def reset_databases():
+    yield
+    await remove_record("20200407142816")
+    await remove_image(
+        ["20200407142816/PM-201-FE-CAM-1.png", "20200407142816/PM-201-FE-CAM-2.png"],
+    )
+    await remove_waveform()
+    if os.path.exists("test.h5"):
+        os.remove("test.h5")
